I have created a basic maze game
My maze game contains skeletons which will bring you back to the start and a score which will increment and decrement based on if you move or not. There are two levels in this game. Level 1 is easy to navigate and has 5 skeletons. Level 2 is harder to navigate and has 6 skeletons. I created this maze game by following a tutorial.